+++ GUI-dev/src/app/pages/admin/admin.component.ts
@@ -1,0 +1,12 @@
+import {Component} from '@angular/core';
+
+@Component({
+  selector: 'ngx-admin',
+  styles: [],
+  template: `<router-outlet></router-outlet>`,
+})
+export class AdminComponent {
+
+  constructor() {
+  }
+}