+++ GUI-dev/src/app/pages/admin/components/importProcessModel/importModelFormBuilder.loader.ts
@@ -1,0 +1,5 @@
+('formBuilder/dist/form-builder.js');
+('formBuilder/dist/form-render.js');
+('style-loader!formBuilder/dist/form-builder.css');
+('style-loader!formBuilder/dist/form-render.css');
+('jquery-ui-bundle');