+++ GUI-dev/src/app/pages/dashboard/components/activeProcessDetail/index.ts
@@ -1,0 +1,2 @@
+export * from './activeProcessDetail.component';
+export * from './businessObjects.component';