+++ GUI-dev/src/app/pages/dashboard/dashboard.component.ts
@@ -5,4 +5,17 @@ import { Component } from '@angular/core';
   templateUrl: './dashboard.component.html',
 })
 export class DashboardComponent {
+
+/*
+  tabs = [
+    {
+      title: 'Active Processes',
+      route: '/activeProcesses',
+    },
+    {
+      title: 'Route tab #2',
+      route: '/activeProcesses',
+    }
+  ];
+*/
 }