+++ GUI-dev/src/app/pages/pages.module.ts
@@ -4,6 +4,9 @@ import { DashboardModule } from './dashboard/dashboard.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
 import {HomeModule} from './home/home.module';
+
+import AdminModule from './admin/admin.module';
+
 import {UserDetailsModule} from './user-details/user-details.module';
 
 const PAGES_COMPONENTS = [
@@ -16,7 +19,9 @@ const PAGES_COMPONENTS = [
     ThemeModule,
     DashboardModule,
     HomeModule,
+    AdminModule,
     UserDetailsModule,
+
 
   ],
   declarations: [