+++ CHANGELOG.md
@@ -4,6 +4,19 @@ This file ONLY contains changes made by the AIM17 dev-team. This means, that thi
 of the forked repository.
 The format is partially based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## 2018-06-04 ([#95](https://github.com/amarbajric/EBUSA-AIM17/pull/95))
+## Added
+- Admin Page
+  - Components in ../pages/admin/components
+    - `activeProcesses`, `eventLogger`, `generateOWL`, `importProcessModel`, `manipulatePNML`, `processModels`, `terminatedProcesses`
+- Components for User dashboard in ../pages/dashboard/components
+  - `activeProcesses`, `activeProcessDetail`, `startableProcesses`, `terminatedProcesses`
+- allProcesses.service
+- evntLogger.service
+## Changed
+- models.ts
+  - added uid for User
+
 ## 2018-06-04 ([#94](https://github.com/amarbajric/EBUSA-AIM17/pull/94))
 ## Added
 - added new Service ProcessStore `at.fhjoanneum.ippr.processstore` running on port 12000